Moved exceptions to domain
