+++ src/main/java/com/emazon/stock/domain/exception/CategoryAlreadyExistsException.java
@@ -1,4 +1,4 @@
-package com.emazon.stock.infrastucture.exception;
+package com.emazon.stock.domain.exception;
 
 public class CategoryAlreadyExistsException  extends RuntimeException{
     public CategoryAlreadyExistsException() {
