+++ src/main/java/com/emazon/stock/domain/spi/ICategoryPersistencePort.java
@@ -6,6 +6,7 @@ import com.emazon.stock.domain.model.PageCustom;
 public interface ICategoryPersistencePort {
 
     void saveCategory(Category category);
+    Boolean findByName(String categoryName);
     PageCustom<Category> getAllCategories(Integer page, Integer size, String sortDirection, String sortBy);
     void updateCategory(Category category);
     void deleteCategory(String categoryName);
