+++ src/main/java/com/emazon/stock/infrastucture/exceptionhandler/ControllerAdvisor.java
@@ -2,8 +2,8 @@ package com.emazon.stock.infrastucture.exceptionhandler;
 import com.emazon.stock.domain.exception.DataConstraintViolationException;
 import com.emazon.stock.domain.exception.MissingAttributeException;
 import com.emazon.stock.domain.exception.PaginationParametersInvalidException;
-import com.emazon.stock.infrastucture.exception.CategoryAlreadyExistsException;
-import com.emazon.stock.infrastucture.exception.CategoryNotFoundException;
+import com.emazon.stock.domain.exception.CategoryAlreadyExistsException;
+import com.emazon.stock.domain.exception.CategoryNotFoundException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
