+++ src/main/java/com/emazon/stock/infrastucture/output/jpa/adapter/CategoryJpaAdapter.java
@@ -3,8 +3,7 @@ package com.emazon.stock.infrastucture.output.jpa.adapter;
 import com.emazon.stock.domain.model.Category;
 import com.emazon.stock.domain.model.PageCustom;
 import com.emazon.stock.domain.spi.ICategoryPersistencePort;
-import com.emazon.stock.infrastucture.exception.CategoryAlreadyExistsException;
-import com.emazon.stock.infrastucture.exception.CategoryNotFoundException;
+import com.emazon.stock.domain.exception.CategoryNotFoundException;
 import com.emazon.stock.infrastucture.output.jpa.entity.CategoryEntity;
 import com.emazon.stock.infrastucture.output.jpa.mapper.CategoryEntityMapper;
 import com.emazon.stock.infrastucture.output.jpa.mapper.PageMapper;
@@ -13,6 +12,8 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
+
+import java.util.Optional;
 
 public class CategoryJpaAdapter implements ICategoryPersistencePort {
 
@@ -28,14 +29,15 @@ public class CategoryJpaAdapter implements ICategoryPersistencePort {
 
     @Override
     public void saveCategory(Category category) {
-        if(categoryRepository.findByName(category.getName()).isPresent()){
-            throw new CategoryAlreadyExistsException();
-        }
-
         CategoryEntity categoryEntity = categoryEntityMapper.toEntity(category);
 
         categoryRepository.save(categoryEntity);
 
+    }
+
+    @Override
+    public Boolean findByName(String categoryName) {
+        return categoryRepository.findByName(categoryName).isPresent();
     }
 
     @Override
@@ -50,11 +52,13 @@ public class CategoryJpaAdapter implements ICategoryPersistencePort {
 
     @Override
     public void updateCategory(Category category) {
-        CategoryEntity categoryEntity = categoryRepository.findByName(category.getName())
-                .orElseThrow(() -> new CategoryNotFoundException("Category not found with name: " + category.getName()));
+        Optional<CategoryEntity> optionalCategoryEntity = categoryRepository.findByName(category.getName());
 
-        categoryEntity.setDescription(category.getDescription());
-        categoryRepository.save(categoryEntity);
+        if (optionalCategoryEntity.isPresent()) {
+            CategoryEntity categoryEntity = optionalCategoryEntity.get();
+            categoryEntity.setDescription(category.getDescription());
+            categoryRepository.save(categoryEntity);
+        }
     }
 
     @Override
